requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
loguru==0.7.2
vk-api==11.9.9
cryptography==41.0.7